--- Lab05/src/pt/c40task/l05wumpus/Componentes.java
@@ -1,21 +1,0 @@
-package pt.c40task.l05wumpus;
-
-public class Componentes {
-	Caverna cave;
-	int i, j;
-	public int getI() {
-		return i;
-	}
-	public void setI(int i) {
-		this.i = i;
-	}
-	public int getJ() {
-		return j;
-	}
-	public void setJ(int j) {
-		this.j = j;
-	}
-	public void conectaCaverna(Caverna cave) {
-		this.cave = cave;
-	}
-}
