+++ Lab05/src/pt/c40task/l05wumpus/Componentes/Buraco.java
@@ -1,0 +1,5 @@
+package pt.c40task.l05wumpus.Componentes;
+
+public class Buraco {
+
+}
