--- Lab05/src/pt/c40task/l05wumpus/Ouro.java
@@ -1,5 +1,0 @@
-package pt.c40task.l05wumpus;
-
-public class Ouro {
-
-}
